backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
astor==0.8.1
black==24.10.0
click==8.3.0
dnspython==2.8.0
fastapi==0.115.14
h11==0.16.0
idna==3.11
injector==0.22.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pydantic==2.12.2
pydantic_core==2.41.4
pymongo==3.12.3
pynest-api==0.4.0
python-dotenv==1.1.1
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.32.1
websockets==15.0.1
